update structure, refactoring
