+++ index.html
@@ -6,6 +6,6 @@
     <link rel="stylesheet" href="src/ui/gui.css">
 </head>
 <body>
-	<script type="module" src="src/index.js"></script>
+	<script type="module" src="src/core/index.js"></script>
 </body>
 </html>
