+++ package.json
@@ -2,7 +2,7 @@
   "name": "smoking-mirror",
   "version": "1.0.0",
   "description": "Another kind of Text Adventure",
-  "main": "index.js",
+  "main": "src/core/index.js",
   "directories": {
     "lib": "lib"
   },
