+++ src/core/EventManager.js
