+++ src/core/Interaction.js
@@ -1,0 +1,27 @@
+
+import Element from "../am/Element.js";
+import Item from "../am/Item.js";
+
+
+const Interaction = class {
+
+    leftClick(element) {
+        if (element instanceof Element) {
+        
+        } else if (element instanceof Item) {
+
+        }
+    }
+
+    rightClick(element) {
+        if (element instanceof Element) {
+        
+        } else if (element instanceof Item) {
+
+        }
+    }
+
+};
+
+
+export default new Interaction();
